Added product CRUD

+++ models/product_model.py
@@ -11,7 +11,7 @@ class Product(Base):
     product_id = Column(Integer, primary_key=True, index=True)
     product_name = Column(String, index=True)
     product_date = Column(DateTime, index=True)
-    product_calorific_value = Column(Float, index=True)
+    product_calorific_value = Column(Integer, index=True)
     user_id = Column(Integer, ForeignKey("users.user_id"))
 
     owner = relationship("User", back_populates="products")
